# Improve speed of to_string for Classes