The prior code would reallocate for every item in the set that is pushed into the buffer. This should make one bigger allocation for the `Vec<&str>` (assuming it uses the size_hint? - worst case it makes a log_n number of allocations to accommodate the str references) and another for joining it into the final `String` with spaces.