@hgzimmerman 2019-12-04
Benchmarks are ran using the nightly  `test::Bencher`.
Tests are ran on Classes constructed from variations on the string:  "class1 class2 class3 ...";`
`old` indicates the present master.
`new` indicates the proposed change.

Tests were ran against sets of 1, 2, 5, and 10 classes as more are unlikely to be used under typical use cases.

---
#### Results
```
test tests::bench_new_1  ... bench:          32 ns/iter (+/- 2)
test tests::bench_new_2  ... bench:          37 ns/iter (+/- 4)
test tests::bench_new_5  ... bench:          50 ns/iter (+/- 5)
test tests::bench_new_10 ... bench:          67 ns/iter (+/- 8)

test tests::bench_old_1  ... bench:          30 ns/iter (+/- 3)
test tests::bench_old_2  ... bench:          46 ns/iter (+/- 4)
test tests::bench_old_5  ... bench:          84 ns/iter (+/- 3)
test tests::bench_old_10 ... bench:         159 ns/iter (+/- 13)
```

#### Analysis
The proposed change is within the margin of error in the 1 class case, while meaningfully faster for inputs larger than 1.